+++ crates/miden-standards/src/account/interface/extension.rs
@@ -16,6 +16,7 @@ use crate::account::components::{
     fungible_token_metadata_library,
     guarded_multisig_library,
     multisig_library,
+    network_account_auth_library,
     network_fungible_faucet_library,
     no_auth_library,
     singlesig_acl_library,
@@ -126,6 +127,10 @@ impl AccountInterfaceExt for AccountInterface {
                 AccountComponentInterface::AuthNoAuth => {
                     component_proc_digests
                         .extend(no_auth_library().mast_forest().procedure_digests());
+                },
+                AccountComponentInterface::AuthNetworkAccount => {
+                    component_proc_digests
+                        .extend(network_account_auth_library().mast_forest().procedure_digests());
                 },
                 AccountComponentInterface::Custom(custom_procs) => {
                     component_proc_digests