+++ crates/miden-testing/tests/auth/mod.rs
@@ -6,3 +6,5 @@ mod multisig;
 mod hybrid_multisig;
 
 mod guarded_multisig;
+
+mod network_account;